requirements.txt
gradio
google-cloud-storage
moviepy
requests